# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LeaseSteals API - US car lease deal tracker
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DealCarType = (typeof DealCarType)[keyof typeof DealCarType];

export const DealCarType = {
  sedan: "sedan",
  suv: "suv",
  truck: "truck",
  ev: "ev",
  coupe: "coupe",
  hatchback: "hatchback",
  van: "van",
} as const;

export interface Deal {
  id: number;
  /** Car manufacturer (e.g. Toyota) */
  make: string;
  /** Car model (e.g. Camry) */
  model: string;
  year: number;
  carType: DealCarType;
  /** MSRP in dollars */
  msrp: number;
  /** Monthly payment in dollars */
  monthlyPayment: number;
  /** Money due at signing (should be 0 for sign and drive) */
  moneyDown: number;
  /** Lease term in months */
  termMonths: number;
  /** Annual mileage limit */
  mileageLimit: number;
  /** Monthly payment as percentage of MSRP (lower is better) */
  dealScore: number;
  /** True if $0 down and deal score < 1.0 */
  isSignAndDrive: boolean;
  /** US region where deal is available (e.g. "National", "SoCal") */
  region: string;
  expiresAt?: string | null;
  imageUrl?: string | null;
  sourceUrl?: string | null;
  /** Trim level (e.g. LE, XLE, Sport) */
  trimLevel?: string | null;
  description?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface DealsListResponse {
  deals: Deal[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export type CreateDealRequestCarType =
  (typeof CreateDealRequestCarType)[keyof typeof CreateDealRequestCarType];

export const CreateDealRequestCarType = {
  sedan: "sedan",
  suv: "suv",
  truck: "truck",
  ev: "ev",
  coupe: "coupe",
  hatchback: "hatchback",
  van: "van",
} as const;

export interface CreateDealRequest {
  make: string;
  model: string;
  year: number;
  carType: CreateDealRequestCarType;
  msrp: number;
  monthlyPayment: number;
  moneyDown?: number;
  termMonths: number;
  mileageLimit: number;
  region: string;
  expiresAt?: string | null;
  imageUrl?: string | null;
  sourceUrl?: string | null;
  trimLevel?: string | null;
  description?: string | null;
}

export type UpdateDealRequestCarType =
  (typeof UpdateDealRequestCarType)[keyof typeof UpdateDealRequestCarType];

export const UpdateDealRequestCarType = {
  sedan: "sedan",
  suv: "suv",
  truck: "truck",
  ev: "ev",
  coupe: "coupe",
  hatchback: "hatchback",
  van: "van",
} as const;

export interface UpdateDealRequest {
  make?: string;
  model?: string;
  year?: number;
  carType?: UpdateDealRequestCarType;
  msrp?: number;
  monthlyPayment?: number;
  moneyDown?: number;
  termMonths?: number;
  mileageLimit?: number;
  region?: string;
  expiresAt?: string | null;
  imageUrl?: string | null;
  sourceUrl?: string | null;
  trimLevel?: string | null;
  description?: string | null;
}

export interface SubscribeRequest {
  email: string;
}

export interface SubscribeResponse {
  message: string;
  email: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type ListDealsParams = {
  /**
   * Filter by car brand/make
   */
  brand?: string;
  /**
   * Filter by car type
   */
  carType?: ListDealsCarType;
  /**
   * Maximum monthly payment filter
   */
  maxMonthly?: number;
  /**
   * Sort field
   */
  sortBy?: ListDealsSortBy;
  /**
   * Sort order
   */
  sortOrder?: ListDealsSortOrder;
  /**
   * @minimum 1
   */
  page?: number;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
};

export type ListDealsCarType =
  (typeof ListDealsCarType)[keyof typeof ListDealsCarType];

export const ListDealsCarType = {
  sedan: "sedan",
  suv: "suv",
  truck: "truck",
  ev: "ev",
  coupe: "coupe",
  hatchback: "hatchback",
  van: "van",
} as const;

export type ListDealsSortBy =
  (typeof ListDealsSortBy)[keyof typeof ListDealsSortBy];

export const ListDealsSortBy = {
  deal_score: "deal_score",
  monthly_payment: "monthly_payment",
  msrp: "msrp",
  created_at: "created_at",
} as const;

export type ListDealsSortOrder =
  (typeof ListDealsSortOrder)[keyof typeof ListDealsSortOrder];

export const ListDealsSortOrder = {
  asc: "asc",
  desc: "desc",
} as const;