# lib/api-zod/src/generated/types/deal.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LeaseSteals API - US car lease deal tracker
 * OpenAPI spec version: 0.1.0
 */
import type { DealCarType } from "./dealCarType";

export interface Deal {
  id: number;
  /** Car manufacturer (e.g. Toyota) */
  make: string;
  /** Car model (e.g. Camry) */
  model: string;
  year: number;
  carType: DealCarType;
  /** MSRP in dollars */
  msrp: number;
  /** Monthly payment in dollars */
  monthlyPayment: number;
  /** Money due at signing (should be 0 for sign and drive) */
  moneyDown: number;
  /** Lease term in months */
  termMonths: number;
  /** Annual mileage limit */
  mileageLimit: number;
  /** Monthly payment as percentage of MSRP (lower is better) */
  dealScore: number;
  /** True if $0 down and deal score < 1.0 */
  isSignAndDrive: boolean;
  /** US region where deal is available (e.g. "National", "SoCal") */
  region: string;
  expiresAt?: Date | null;
  imageUrl?: string | null;
  sourceUrl?: string | null;
  /** Trim level (e.g. LE, XLE, Sport) */
  trimLevel?: string | null;
  description?: string | null;
  createdAt: Date;
  updatedAt: Date;
}