# lib/api-zod/src/generated/types/dealsListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LeaseSteals API - US car lease deal tracker
 * OpenAPI spec version: 0.1.0
 */
import type { Deal } from "./deal";

export interface DealsListResponse {
  deals: Deal[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}