# lib/api-zod/src/generated/types/subscribeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LeaseSteals API - US car lease deal tracker
 * OpenAPI spec version: 0.1.0
 */

export interface SubscribeRequest {
  email: string;
}