# lib/api-zod/src/generated/types/updateDealRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LeaseSteals API - US car lease deal tracker
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateDealRequestCarType } from "./updateDealRequestCarType";

export interface UpdateDealRequest {
  make?: string;
  model?: string;
  year?: number;
  carType?: UpdateDealRequestCarType;
  msrp?: number;
  monthlyPayment?: number;
  moneyDown?: number;
  termMonths?: number;
  mileageLimit?: number;
  region?: string;
  expiresAt?: Date | null;
  imageUrl?: string | null;
  sourceUrl?: string | null;
  trimLevel?: string | null;
  description?: string | null;
}